backend/etrosa/pyproject.toml
[tool.poetry]
name = "etrosa"
version = "0.1.0"
description = ""
authors = [

]
maintainers = [

]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.89.1"
uvicorn = { version = "^0.20.0", extras = ["standard"] }
pydantic = {extras = ["email"], version = "^1.10.5"}
yarl = "^1.8.2"
ujson = "^5.7.0"
tortoise-orm = "^0.19.2"
aerich = "^0.7.1"
aiomysql = "^0.1.1"
mysqlclient = "^2.1.1"
cryptography = "^39.0.0"
httptools = "^0.5.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.5"

[tool.poetry.dev-dependencies]
pytest = "^7.2.1"
flake8 = "~4.0.1"
mypy = "^0.991"
isort = "^5.11.4"
pre-commit = "^3.0.1"
wemake-python-styleguide = "^0.17.0"
black = "^22.12.0"
autoflake = "^1.6.1"
pytest-cov = "^4.0.0"
anyio = "^3.6.2"
pytest-env = "^0.8.1"
asynctest = "^0.13.0"
nest-asyncio = "^1.5.6"
httpx = "^0.23.3"

[tool.isort]
profile = "black"
multi_line_output = 3
src_paths = ["etrosa",]

[tool.mypy]
strict = true
ignore_missing_imports = true
allow_subclassing_any = true
allow_untyped_calls = true
pretty = true
show_error_codes = true
implicit_reexport = true
allow_untyped_decorators = true
warn_unused_ignores = false
warn_return_any = false
namespace_packages = true

[tool.pytest.ini_options]
filterwarnings = [
    "error",
    "ignore::DeprecationWarning",
    "ignore:.*unclosed.*:ResourceWarning",
]
env = [
    "ETROSA_DB_BASE=etrosa_test",
]
[tool.aerich]
tortoise_orm = "etrosa.db.config.TORTOISE_CONFIG"
location = "./etrosa/db/migrations"
src_folder = "./."
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
